Configure H2 with large window sizes for backup workloads
The default hyper H2 settings have small flow control windows:
- 65KB per stream window
- 1MB connection window

This causes issues with PBS backup protocol which transfers chunks
up to 4MB. The connection would fail with "stream no longer needed"
when flow control blocked the upload.

Fix by configuring H2 builder with:
- 32MB per stream window
- 64MB connection window
- 100 max concurrent streams
- 16MB max frame size

Also add error logging for H2 connection failures to aid debugging.

+++ crates/pbs-server/src/system_info.rs
@@ -353,10 +353,10 @@ fn read_storage_info(path: &Path) -> Option<StorageInfo> {
         if result != 0 {
             return None;
         }
-        let block_size = vfs.f_frsize;
-        let total = vfs.f_blocks * block_size;
-        let free = vfs.f_bfree * block_size;
-        let avail = vfs.f_bavail * block_size;
+        let block_size = vfs.f_frsize as u64;
+        let total = (vfs.f_blocks as u64) * block_size;
+        let free = (vfs.f_bfree as u64) * block_size;
+        let avail = (vfs.f_bavail as u64) * block_size;
         let used = total.saturating_sub(free);
         Some(StorageInfo { total, used, avail })
     }
